美化06_utilities.c文件：改进变量命名规范和函数注释
- 将ProcessFloatDataValidationAndConversion函数重命名为ValidateAndConvertFloatData
- 改进变量命名：InputFloatValue改为OriginalFloatValue，ResourceIndex改为ResourceEntryPointer等
- 完善函数文档注释，添加更详细的功能说明
- 优化代码注释，提高可读性

+++ TaleWorlds.Native/src/05_networking.c
@@ -806,7 +806,7 @@ NetworkHandle InitializeConnectionHandler(void)
  * @param PacketData 数据包数据
  * @return 处理结果句柄，0表示成功，其他值表示错误码
  */
-NetworkHandle ProcessNetworkConnectionData(long long *ConnectionContext, int PacketData)
+NetworkHandle ProcessNetworkConnectionData(int64_t *ConnectionContext, int32_t PacketData)
 {
   NetworkStatus *NetworkConnectionStatusData;
   int ConnectionCount;
